removed junk files
