--- .idea/compiler.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <bytecodeTargetLevel>
-      <module name="HRVLib_main" target="1.8" />
-      <module name="HRVLib_test" target="1.8" />
-    </bytecodeTargetLevel>
-  </component>
-</project>
