--- .idea/libraries/Gradle__com_google_guava_guava_20_0.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: com.google.guava:guava:20.0">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.google.guava/guava/20.0/89507701249388e1ed5ddcf8c41f4ce1be7831ef/guava-20.0.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.google.guava/guava/20.0/9c8493c7991464839b612d7547d6c263adf08f75/guava-20.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
