--- .idea/libraries/Gradle__org_apache_commons_commons_math3_3_6_1.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: org.apache.commons:commons-math3:3.6.1">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.apache.commons/commons-math3/3.6.1/e4ba98f1d4b3c80ec46392f25e094a6a2e58fcbf/commons-math3-3.6.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.apache.commons/commons-math3/3.6.1/8fab23986ea8886af34818daf32a718e81dc98ba/commons-math3-3.6.1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
