--- .idea/modules.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/HRVLib.iml" filepath="$PROJECT_DIR$/.idea/modules/HRVLib.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/HRVLib_main.iml" filepath="$PROJECT_DIR$/.idea/modules/HRVLib_main.iml" group="HRVLib" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/HRVLib_test.iml" filepath="$PROJECT_DIR$/.idea/modules/HRVLib_test.iml" group="HRVLib" />
-    </modules>
-  </component>
-</project>
